requirements.minimal.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pypdf==3.17.0
jinja2==3.1.2
requests==2.31.0